Minor Upgrades (#173)
* chore: update to latest Rust stable

* chore: run clippy

* chore: upgrade minor package versions

This does not fix issue 171.

* fixup! chore: run clippy

* fixup! chore: run clippy

* fixup! chore: run clippy

+++ src/round_constants.rs
@@ -80,7 +80,7 @@ pub fn generate_constants<F: PrimeField>(
 }
 
 fn append_bits<T: Into<u128>>(vec: &mut Vec<bool>, n: usize, from: T) {
-    let val = from.into() as u128;
+    let val = from.into();
     for i in (0..n).rev() {
         vec.push((val >> i) & 1 != 0);
     }
